dedupe setPattern in draw_fill/extrude code
